+++ ultracache/tests/settings/110.py
@@ -30,8 +30,6 @@ CACHES = {
     }
 }
 
-SITE_ID = 1
-
 SECRET_KEY = "SECRET_KEY"
 
 TEMPLATE_DIRS = (os.path.realpath(os.path.dirname(__file__)) + "/ultracache/tests/templates/",)